forge fmt

+++ test/TVector.t.sol
@@ -9,19 +9,13 @@ contract TVectorTest is Test {
     using TVector for TVector.Vector;
 
     function testNewVector() public {
-        TVector.Vector memory vector = TVector.newVector(
-            AllocatorFactory.AllocatorType.Transient,
-            4
-        );
+        TVector.Vector memory vector = TVector.newVector(AllocatorFactory.AllocatorType.Transient, 4);
         assertEq(vector.length(), 0);
         assertEq(vector.capacity, 4);
     }
 
     function testPushAndAt() public {
-        TVector.Vector memory vector = TVector.newVector(
-            AllocatorFactory.AllocatorType.Transient,
-            4
-        );
+        TVector.Vector memory vector = TVector.newVector(AllocatorFactory.AllocatorType.Transient, 4);
 
         vector.push(1);
         vector.push(2);
@@ -34,10 +28,7 @@ contract TVectorTest is Test {
     }
 
     function testPop() public {
-        TVector.Vector memory vector = TVector.newVector(
-            AllocatorFactory.AllocatorType.Transient,
-            4
-        );
+        TVector.Vector memory vector = TVector.newVector(AllocatorFactory.AllocatorType.Transient, 4);
 
         vector.push(1);
         vector.push(2);
@@ -49,30 +40,21 @@ contract TVectorTest is Test {
     }
 
     function testPopEmpty() public {
-        TVector.Vector memory vector = TVector.newVector(
-            AllocatorFactory.AllocatorType.Transient,
-            4
-        );
+        TVector.Vector memory vector = TVector.newVector(AllocatorFactory.AllocatorType.Transient, 4);
 
         vm.expectRevert(TVector.TVectorEmpty.selector);
         vector.pop();
     }
 
     function testAtOutOfBounds() public {
-        TVector.Vector memory vector = TVector.newVector(
-            AllocatorFactory.AllocatorType.Transient,
-            4
-        );
+        TVector.Vector memory vector = TVector.newVector(AllocatorFactory.AllocatorType.Transient, 4);
 
         vm.expectRevert(TVector.TVectorOutOfBounds.selector);
         vector.at(0);
     }
 
     function testAutoResize() public {
-        TVector.Vector memory vector = TVector.newVector(
-            AllocatorFactory.AllocatorType.Transient,
-            2
-        );
+        TVector.Vector memory vector = TVector.newVector(AllocatorFactory.AllocatorType.Transient, 2);
 
         vector.push(1);
         vector.push(2);
@@ -85,10 +67,7 @@ contract TVectorTest is Test {
     }
 
     function testMemoryAllocator() public {
-        TVector.Vector memory vector = TVector.newVector(
-            AllocatorFactory.AllocatorType.Memory,
-            4
-        );
+        TVector.Vector memory vector = TVector.newVector(AllocatorFactory.AllocatorType.Memory, 4);
 
         vector.push(1);
         vector.push(2);
@@ -101,10 +80,7 @@ contract TVectorTest is Test {
     }
 
     function testMemoryAllocatorResize() public {
-        TVector.Vector memory vector = TVector.newVector(
-            AllocatorFactory.AllocatorType.Memory,
-            2
-        );
+        TVector.Vector memory vector = TVector.newVector(AllocatorFactory.AllocatorType.Memory, 2);
 
         vector.push(1);
         vector.push(2);
@@ -123,21 +99,15 @@ contract TVectorTest is Test {
 
     function testCapacityTooLarge() public {
         vm.expectRevert(TVector.TVectorCapacityTooLarge.selector);
-        TVector.newVector(
-            AllocatorFactory.AllocatorType.Transient,
-            type(uint256).max
-        );
+        TVector.newVector(AllocatorFactory.AllocatorType.Transient, type(uint256).max);
     }
 
     function testFuzzPushAndAt(uint256[] memory values) public {
         vm.assume(values.length > 0 && values.length <= 100);
 
-        TVector.Vector memory vector = TVector.newVector(
-            AllocatorFactory.AllocatorType.Transient,
-            4
-        );
+        TVector.Vector memory vector = TVector.newVector(AllocatorFactory.AllocatorType.Transient, 4);
 
-        for (uint i = 0; i < values.length; i++) {
+        for (uint256 i = 0; i < values.length; i++) {
             vector.push(values[i]);
             assertEq(vector.at(i), values[i]);
         }
